Rename runReports to getReportRunner to more accurately reflect what's going on
