+++ lib/bot/getReportRunner.js
