+++ lib/bot/index.js
@@ -5,5 +5,5 @@ module.exports = {
   getUserStandupInfo: require('./getUserStandupInfo'),
   joinChannel: require('./joinChannel'),
   replyToBadStandup: require('./replyToBadStandup'),
-  runReports: require('./runReports')
+  getReportRunner: require('./getReportRunner')
 };
